services/prism-console-api/pyproject.toml
[tool.poetry]
name = "prism-console-api"
version = "0.1.0"
description = "FastAPI backend for BlackRoad Prism Console"
authors = ["BlackRoad Ops <ops@blackroad.io>"]
packages = [{ include = "prism", from = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
sqlmodel = "^0.0.14"
httpx = { extras = ["http2"], version = "^0.27.0" }
structlog = "^24.1.0"
prometheus-client = "^0.19.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
jsonschema = "^4.21.1"
itsdangerous = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^5.0.0"
httpx = { extras = ["http2"], version = "^0.27.0" }
respx = "^0.21.1"
aioresponses = "^0.7.6"
ruff = "^0.4.1"
uvloop = { version = "^0.19.0", optional = true }
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
ignore = ["E203"]

[tool.coverage.run]
branch = true
source = ["src/prism"]

[tool.coverage.report]
omit = ["tests/*"]
